requirements.txt
Babel==0.9.6
Flask==0.9
Flask-Babel==0.8
Flask-Login==0.1.3
Flask-Mail==0.7.3
Flask-OAuth==0.12
Flask-OpenID==1.1
Flask-Principal==0.3.3
Flask-SQLAlchemy==0.16
Flask-Security==1.5.0
Flask-Social==1.5.1
Flask-WTF==0.8
Flask-WhooshAlchemy==0.55a
Jinja2==2.6
MySQL-python==1.2.4c1
SQLAlchemy==0.7.9
Tempita==0.5.1
WTForms==1.0.2
Werkzeug==0.8.3
Whoosh==2.4.1
argparse==1.2.1
blinker==1.2
decorator==3.4.0
facebook-sdk==0.4.0
## FIXME: could not find svn URL in dependency_links for this package:
flup==1.0.3.dev-20110405
httplib2==0.7.7
itsdangerous==0.17
oauth2==1.5.211
passlib==1.6.1
psycopg2==2.4.5
python-openid==2.2.5
python-twitter==0.8.2
pytz==2012h
simplejson==2.6.2
speaklater==1.3
sqlalchemy-migrate==0.7.2
wsgiref==0.1.2
